hw8 corrected

+++ Test2/presentation/src/main/java/com/runkevich8/gmail/test/homework/hw8/PublishContract.java
@@ -4,5 +4,5 @@ package com.runkevich8.gmail.test.homework.hw8;
 import io.reactivex.Observable;
 
 public interface PublishContract {
-    Observable<Integer> getObservable();
+    Observable<Long> getObservable();
 }
